requirements.txt
av==14.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
decorator==5.2.1
Flask==3.1.0
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.1.2
numpy==2.2.4
packaging==24.2
pillow==10.4.0
pillow_heif==0.22.0
proglog==0.1.11
pygame==2.6.1
python-dotenv==1.1.0
tqdm==4.67.1
Werkzeug==3.1.3
